requirements.txt
python==3.6.13
pandas==1.1.5
numpy==1.19.2
scikit-learn==0.24.2
tqdm==4.63.0
pyyaml==6.0
tensorflow==1.12.0
